Split out functions for construction or Renderer types into ShaderRendererUtil. Removed the serialization of buffers code into test-render
